Support arbitrary signature parameters.
The initial draft of this document rejected any `Signature-Input` member
whose parameters included unknown names. WICG/signature-based-sri#38
lands on a different conclusion, accepting any parameter that's provided
and including it in the serialization of the signature base. That should
alleviate potential deployment concerns that have been raised, and will
hopefully do so in a way that improves forward compatibility.

Closes #38.